A bunch more mypy errors fixed
